Progress towards Synced music/beats
I've made a gameObject in the tutorial level: LevelManager. It is a parent of leader/player manager and tutorial level music (although I ended up moving all music functionality into Level Manager itself)

This gameObject, in coordination with GameHandler, has set up difinitive timing with the music. It currently does not do anything other than
1) Manage pausing of the music
2) Output a debug statement with (ideally) percise timing of where our leader/player sequences should go.

It has NOT fixed the leader/player manager sync, to do that I believe we should call the leader/player manager functions from within the LevelManager script, using the function that is synced with the music to our advantage.