Draft of Modular Rhythm Design
A new GameHandler, defined in WorkshopScene_Carlos, impliments a new rhymth handling system. The idea is that, just from the inspector, the user can asing:
1) Music and bpm
2) Up to 4 keys which the player will be using
3) A complete map of when and what keys the player must input, based on the measure, beat, and quarternote of the song.
And then the New_GameHandler will play the song and request the provided user inputs exactly when and where it was told. The player will have a quaternote window to hit these the specified key.

I highly recommend looking first at Define_BeatMap.cs first to understand how it tracks what keys (1, 2, 3, or 4) the player must press and when. Then look at Manage_Game.cs to understand how to work with the handler and possibly add a scoring system.
But here's the best part: You can do art/other work almost seperatly from this gameobject. Anytime you want to find the current measure/beat/note of the song, use the code in Manage_Game.cs lines 66-70. If you want to know the map of what the player's input should be, retreive the beatMap from the New_GameHandler. These two things give all the information needed to know where in the song we are and what player input is coming up in the future.

Contact Carlos for questions