System for Modularly defining beatmaps
There are officially a few ways to do this, but if you have a script that uses a beatmap in a special way and you need different definitions of a beatmap depending on the level in that same script, use a public member of the script MakebeatMap and call its only function to get your beatmap. Then in another script which is a child of MakebeatMap, define the beatmap you want. Example in WorkshopScene_Carlos